test: split test_tools_ci.py into 4 per-tool files
Split the 776-line test_tools_ci.py (33 module-level test functions + 2
classes: TestWaitForCiTiming, TestWaitForMergeQueueTiming) into:

- test_tools_ci_gate.py (~80 LoC): 6 functions covering GATED_TOOLS / UNGATED_TOOLS
  membership and gate-closure behavior for wait_for_ci / get_ci_status /
  wait_for_merge_queue
- test_tools_ci_wait_for_ci.py (~360 LoC): 15 functions + class TestWaitForCiTiming
  covering the wait_for_ci tool surface (success/failure, head_sha inference,
  event/lookback forwarding, remote_url, exception handling, timing)
- test_tools_ci_get_ci_status.py (~50 LoC): 3 functions covering the get_ci_status
  tool surface (missing branch/run_id, no watcher, workflow propagation)
- test_tools_ci_wait_for_merge_queue.py (~280 LoC): 9 functions + class
  TestWaitForMergeQueueTiming covering the wait_for_merge_queue tool surface
  (delegation, repo inference, remote_url, exception handling,
  PRState round-trip exhaustiveness, timing)

Per-tool naming follows the established test_tools_<tool>.py convention.
Coexistence with test_tools_ci_enqueue.py and test_tools_ci_merge_state.py is
unaffected — those files cover different production tools (enqueue_pr,
check_repo_merge_state).

The actual split is disjoint and complete (33 functions total). The plan
estimate of 26 was a count discrepancy; the plan said ~15 for wait_for_ci
which matches the actual count.

Mechanical lift: no production code changes, no behavior changes, no test
assertion changes. Each file carries pytestmark = [layer("server"), small].

Refs: #4610
